Sort some tests, discover that allowAppImports makes things hard
