+++ test-app/app/allow-app-imports.js
@@ -1,0 +1,1 @@
+export const globs = import.meta.glob('./allow-app-imports/*');
