+++ test-app/app/allow-app-imports/a.js
@@ -1,0 +1,1 @@
+export const a = 'a1';
