+++ test-app/app/allow-app-imports/b.js
@@ -1,0 +1,1 @@
+export const b = 'b1';
