+++ test-app/app/allow-app-imports/c.js
@@ -1,0 +1,1 @@
+export const c = 'c1';
