+++ test-app/ember-cli-build.js
@@ -6,6 +6,7 @@ module.exports = function (defaults) {
   let app = new EmberApp(defaults, {
     autoImport: {
       watchDependencies: ['ember-classic-import-meta-glob'],
+      allowAppImports: ['allow-app-imports/*'],
     },
   });
