+++ test-app/tests/glob-test/the-test.js
@@ -1,12 +1,10 @@
-import { module, test } from 'qunit';
-import { importMetaGlob } from 'ember-classic-import-meta-glob';
+import { module, test, skip } from 'qunit';
 import { globs as appTreeGlob } from 'test-app/glob-imports';
+import { globs as allowAppImports } from 'test-app/allow-app-imports';
 
 const testTreeGlob = import.meta.glob('./from-tests/**', { eager: true });
 
-const HERE = 'test-app/tests/glob-test/the-test';
-
-module('Glob tests', function () {
+module('from the app tree', function () {
   test('works from app', function (assert) {
     let keys = Object.keys(appTreeGlob);
     assert.deepEqual(keys, ['./from-app/a', './from-app/b', './from-app/c']);
@@ -15,6 +13,29 @@ module('Glob tests', function () {
     assert.strictEqual(appTreeGlob['./from-app/c'].c, 'c1');
   });
 
+  /**
+   * Can't test this atm, because await import()
+   * in our runtime code is not allowed to receive a
+   * dynamic string.
+   *
+   * The interface of the returned object should match
+   * the real implementation in vite/babel/etc
+   */
+  skip('default (async) w/ app-imports', async (assert) => {
+    let keys = Object.keys(allowAppImports);
+    assert.deepEqual(keys, [
+      './allow-app-imports/a',
+      './allow-app-imports/b',
+      './allow-app-imports/c',
+    ]);
+
+    assert.strictEqual((await allowAppImports['./from-tests/a']()).a, 'a1');
+    assert.strictEqual((await allowAppImports['./from-tests/b']()).b, 'b1');
+    assert.strictEqual((await allowAppImports['./from-tests/c']()).c, 'c1');
+  });
+});
+
+module('from the test tree', function () {
   test('works from tests', function (assert) {
     let keys = Object.keys(testTreeGlob);
     assert.deepEqual(keys, [
@@ -26,7 +47,9 @@ module('Glob tests', function () {
     assert.strictEqual(appTreeGlob['./from-app/b'].b, 'b1');
     assert.strictEqual(appTreeGlob['./from-app/c'].c, 'c1');
   });
+});
 
+module('from this file', function () {
   test('with no matches', function (assert) {
     let result = import.meta.glob('./does/not/exist', { eager: true });
 
@@ -47,77 +70,5 @@ module('Glob tests', function () {
     assert.strictEqual((await result['./from-tests/a']()).a, 'a1');
     assert.strictEqual((await result['./from-tests/b']()).b, 'b1');
     assert.strictEqual((await result['./from-tests/c']()).c, 'c1');
-  });
-
-  module('underlying runtime', function () {
-    test('extensions are stripped and the same as without using extensions', (assert) => {
-      let a = importMetaGlob('./from-tests/**/*.js', { eager: true }, HERE);
-      let b = importMetaGlob('./from-tests/**/*', { eager: true }, HERE);
-
-      assert.deepEqual(Object.keys(a), Object.keys(b));
-      assert.deepEqual(Object.values(a)[0].a, Object.values(b)[0].a);
-      assert.deepEqual(Object.values(a)[1].b, Object.values(b)[1].b);
-      assert.deepEqual(Object.values(a)[2].c, Object.values(b)[2].c);
-    });
-
-    module('errors', function () {
-      module('glob', function () {
-        test('errors when trying to escape the app', function (assert) {
-          assert.throws(() => {
-            importMetaGlob('../../../something', { eager: true }, HERE);
-          }, /Cannot have a path that escapes the app/);
-        });
-
-        test('errors when a sibling path tries to escape the app', function (assert) {
-          assert.throws(() => {
-            importMetaGlob(
-              './from-tests/../../../../something',
-              { eager: true },
-              HERE,
-            );
-          }, /Cannot have a path that escapes the app/);
-        });
-
-        test('invalid glob', function (assert) {
-          assert.throws(() => {
-            importMetaGlob('**/*', HERE);
-          }, /The glob pattern must be a relative path starting with either/);
-        });
-      });
-
-      module('options', function () {
-        test('errors with incorrect options', function (assert) {
-          assert.throws(() => {
-            importMetaGlob('./**/*', true, HERE);
-          }, /the second argument to import.meta.glob must be an object/);
-        });
-
-        test('when passing options, cannot be empty', function (assert) {
-          assert.throws(() => {
-            importMetaGlob('./**/*', {}, HERE);
-          }, /the only supported option is 'eager'/);
-        });
-
-        test('when passing options, only eager is allowed', function (assert) {
-          assert.throws(() => {
-            importMetaGlob('./**/*', { boop: 1 }, HERE);
-          }, /the only supported option is 'eager'/);
-        });
-      });
-
-      module('modulePath', function () {
-        test('errors with no modulePath', function (assert) {
-          assert.throws(() => {
-            importMetaGlob('./**/*', { eager: true });
-          }, /the third argument to import.meta.glob must be passed and be the module path/);
-        });
-
-        test(`errors with wrong path (can't be the entrypoint)`, function (assert) {
-          assert.throws(() => {
-            importMetaGlob('../../../something', { eager: true }, 'test-app');
-          }, /not a valid path/);
-        });
-      });
-    });
   });
 });
